chore: remove dead search code
This hasn't been used for a while and was getting in the way of some of the
other dead code I've been pruning.

GraphqlCursor from search-protocol _was_ being used in a few places, so I've
extracted it into its own crate.
