+++ Cargo.toml
@@ -114,6 +114,7 @@ gateway-types = { path = "engine/crates/gateway-types" }
 gateway-v2 = { path = "engine/crates/gateway-v2" }
 graph-entities = { path = "engine/crates/graph-entities" }
 graphql-composition = { path = "engine/crates/composition" }
+graphql-cursor = { path = "engine/crates/graphql-cursor" }
 graphql-extensions = { path = "engine/crates/graphql-extensions" }
 graphql-mocks = { path = "engine/crates/graphql-mocks" }
 jwt-verifier = { path = "engine/crates/jwt-verifier" }
@@ -124,7 +125,6 @@ parser-postgres = { path = "engine/crates/parser-postgres" }
 runtime-local = { path = "engine/crates/runtime-local" }
 runtime-noop = { path = "engine/crates/runtime-noop" }
 runtime = { path = "engine/crates/runtime" }
-search-protocol = { path = "engine/crates/search-protocol" }
 worker-env = { path = "engine/crates/worker-env" }
 
 [profile.bench]
