+++ engine/crates/graphql-cursor/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "graphql-cursor"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+repository.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde.workspace = true
+serde_with.workspace = true
+
+[lints]
+workspace = true
