+++ engine/crates/runtime-local/src/lib.rs
@@ -4,7 +4,6 @@ mod fetch;
 mod kv;
 mod log;
 mod pg;
-pub mod search;
 mod ufd_invoker;
 
 pub use bridge::Bridge;
@@ -12,7 +11,6 @@ pub use cache::InMemoryCache;
 pub use fetch::NativeFetcher;
 pub use kv::*;
 pub use pg::LocalPgTransportFactory;
-pub use search::LocalSearchEngine;
 pub use ufd_invoker::UdfInvokerImpl;
 
 pub use crate::log::LogEventReceiverImpl;
