+++ engine/crates/runtime/Cargo.toml
@@ -29,7 +29,6 @@ tracing.workspace = true
 tracing-futures.workspace = true
 url.workspace = true
 
-search-protocol = { path = "../search-protocol" }
 postgres-connector-types = { path = "../postgres-connector-types" }
 common-types.workspace = true
