+++ engine/crates/runtime/src/lib.rs
@@ -1,10 +1,10 @@
 pub mod cache;
 pub mod context;
+pub mod cursor;
 pub mod fetch;
 pub mod kv;
 pub mod log;
 pub mod pg;
-pub mod search;
 pub mod udf;
 
 pub use context::Context;
