--- engine/crates/search-protocol/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "search-protocol"
-authors = ["Grafbase"]
-description = "Types related to the Grafbase platform search functionality"
-version = "0.1.0"
-edition = "2021"
-license = "MPL-2.0"
-homepage = "https://grafbase.com"
-repository = "https://github.com/grafbase/grafbase"
-keywords = ["types", "search", "grafbase"]
-
-[lints]
-workspace = true
-
-[dependencies]
-serde.workspace = true
-serde_with = { workspace = true, features = ["base64"] }
-thiserror = "1"
-chrono = { version = "0.4", features = ["serde"] }
-ulid = { workspace = true, features = ["serde"] }
-strum = { workspace = true, features = ["derive"] }
-
-[dev-dependencies]
-serde_json.workspace = true
-rstest.workspace = true
