--- engine/crates/search-protocol/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod config;
-pub mod query;
-mod serde_utils;
