--- engine/crates/search-protocol/src/serde_utils/mod.rs
@@ -1,1 +1,0 @@
-pub mod naive_date;
