--- engine/crates/search-protocol/src/serde_utils/naive_date.rs
@@ -1,34 +1,0 @@
-use chrono::NaiveDate;
-use serde::{self, Deserialize, Deserializer, Serializer};
-
-const FORMAT: &str = "%Y-%m-%d";
-
-// The signature of a serialize_with function must follow the pattern:
-//
-//    fn serialize<S>(&T, S) -> Result<S::Ok, S::Error>
-//    where
-//        S: Serializer
-//
-// although it may also be generic over the input types T.
-#[allow(clippy::trivially_copy_pass_by_ref)]
-pub fn serialize<S>(date: &NaiveDate, serializer: S) -> Result<S::Ok, S::Error>
-where
-    S: Serializer,
-{
-    serializer.serialize_str(&date.format(FORMAT).to_string())
-}
-
-// The signature of a deserialize_with function must follow the pattern:
-//
-//    fn deserialize<'de, D>(D) -> Result<T, D::Error>
-//    where
-//        D: Deserializer<'de>
-//
-// although it may also be generic over the output types T.
-pub fn deserialize<'de, D>(deserializer: D) -> Result<NaiveDate, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let s = String::deserialize(deserializer)?;
-    NaiveDate::parse_from_str(&s, FORMAT).map_err(serde::de::Error::custom)
-}
